# Conflicts:
#	Assets/Resources/Fonts/GL-StellaMystica SDF.asset
#	Assets/Resources/Fonts/Pingsan SDF.asset
#	Assets/TextMesh Pro/Examples & Extras/Resources/Fonts & Materials/Anton SDF.asset
#	Assets/TextMesh Pro/Examples & Extras/Resources/Fonts & Materials/Bangers SDF.asset
#	Assets/TextMesh Pro/Examples & Extras/Resources/Fonts & Materials/Roboto-Bold SDF.asset

+++ Assets/Scripts/Game/AI/AIs/PlayerAI/Tasks/Movement/TaskMove.cs
@@ -65,7 +65,7 @@ public class TaskMove : BaseNode
         }
 
         // apply movement (-1) because of team effect
-        m_Movement.SetMovement((-1) * m_CurrentMoveX);
+        m_Movement.SetMovement((sbyte)((-1) * m_CurrentMoveX));
 
         ErrorHandler.Log("TaskMove - " + m_State, ELogTag.AITaskMove);
